Change: Move useGmp hook to hooks directory
