+++ src/web/components/link/manuallink.jsx
@@ -3,7 +3,6 @@
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
-
 import React from 'react';
 
 import {getLocale} from 'gmp/locale/lang';
@@ -11,7 +10,7 @@ import {getLocale} from 'gmp/locale/lang';
 import {isDefined} from 'gmp/utils/identity';
 
 import PropTypes from 'web/utils/proptypes';
-import useGmp from 'web/utils/useGmp';
+import useGmp from 'web/hooks/useGmp';
 
 import BlankLink from './blanklink';
