+++ src/web/components/provider/licenseprovider.jsx
@@ -3,10 +3,9 @@
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
-
 import React, {useEffect, useState} from 'react';
 
-import useGmp from 'web/utils/useGmp';
+import useGmp from 'web/hooks/useGmp';
 
 export const LicenseContext = React.createContext({});
