+++ src/web/entities/BulkTags.jsx
@@ -13,7 +13,7 @@ import {apiType, getEntityType, typeName} from 'gmp/utils/entitytype';
 import PropTypes from 'web/utils/proptypes';
 import SelectionType from 'web/utils/selectiontype';
 
-import useGmp from 'web/utils/useGmp';
+import useGmp from 'web/hooks/useGmp';
 import useUserSessionTimeout from 'web/hooks/useUserSessionTimeout';
 
 import TagDialog from 'web/pages/tags/dialog';
