+++ src/web/pages/portlists/listpage.jsx
@@ -14,7 +14,7 @@ import {isDefined, hasValue} from 'gmp/utils/identity';
 import useCapabilities from 'web/hooks/useCapabilities';
 import useUserSessionTimeout from 'web/hooks/useUserSessionTimeout';
 
-import useGmp from 'web/utils/useGmp';
+import useGmp from 'web/hooks/useGmp';
 import usePageFilter from 'web/hooks/usePageFilter';
 import useShallowEqualSelector from 'web/hooks/useShallowEqualSelector';
 import useReload from 'web/hooks/useReload';
